Add state 'STATE_STOP' where 'pose' and 'delta_pose' msgs are ignored
Added a state that is entered when service '/teleop/stop' is called and
left when '/teleop/start' is called. When in state 'STATE_STOP', all 'pose'
and 'delta_pose' messages are ignored such that 'STATE_IDLE' and thus
'STATE_TELEOP' can only be entered after '/teleop/start' is called again.